refactor(sp): remove legacy batch.ts + fieldCache.ts duplicates
- Move clearFieldsCacheFor/clearAllFieldsCache from fieldCache.ts → helpers.ts
- Redirect spClient.ts imports: batch.ts → spBatch.ts, fieldCache.ts → helpers.ts + spSchema.ts
- Fix buildBatchPayload call to pass buildItemPath (4-arg signature from spBatch.ts)
- Add buildItemPath to spClient.ts imports from helpers.ts
- Update index.ts barrel to export cache clear utilities from helpers.ts
- Delete redundant batch.ts (duplicate of spBatch.ts)
- Delete redundant fieldCache.ts (all functions now in helpers.ts/spSchema.ts)
- Zero TypeScript errors
