+++ src/lib/sp/index.ts
@@ -26,8 +26,7 @@ export {
     DEFAULT_LIST_TEMPLATE, DEFAULT_STAFF_LIST_TITLE, DEFAULT_USERS_LIST_TITLE,
     FIELDS_CACHE_TTL_MS, STAFF_BASE_FIELDS, STAFF_OPTIONAL_FIELDS,
     USERS_BASE_FIELDS, USERS_OPTIONAL_FIELDS,
-    buildItemPath, buildListItemsPath, buildSelectFields, clampTop,
-    classifyRetry, extractMissingField, getMissingSet, makeFieldsCacheKey,
+    buildItemPath, buildListItemsPath, buildSelectFields, clampTop, classifyRetry, clearAllFieldsCache, clearFieldsCacheFor, extractMissingField, getMissingSet, makeFieldsCacheKey,
     markOptionalMissing, normalizeGuidCandidate, nowMs, raiseHttpError,
     readErrorPayload, resetMissingOptionalFieldsCache, resolveListPath,
     resolveStaffListIdentifier, safeJsonParse, safeJsonStringify,
